requirement.txt
Django
psycopg2
gunicorn
django-import-export
django-taggit
libpq-dev
postgresql
postgresql-contrib
psycopg2